big update by ardila
Added and arranged dependencies and repositories, fixed 100 errors in different classes and activated Maven, Junit and Mokito to perform unit tests.

+++ demo/src/main/java/co/edu/udes/activity/backend/demo/controllers/AcademicRecordController.java
@@ -21,7 +21,7 @@ public class AcademicRecordController {
     }
 
     @GetMapping("/{id}")
-    public Optional<AcademicRecord> getAcademicRecordById(@PathVariable int id) {
+    public Optional<AcademicRecord> getAcademicRecordById(@PathVariable long id) {
         return academicRecordService.getAcademicRecordById(id);
     }
 
@@ -31,12 +31,12 @@ public class AcademicRecordController {
     }
 
     @PutMapping("/{id}")
-    public AcademicRecord updateAcademicRecord(@PathVariable int id, @RequestBody AcademicRecord updatedAcademicRecord) {
+    public AcademicRecord updateAcademicRecord(@PathVariable long id, @RequestBody AcademicRecord updatedAcademicRecord) {
         return academicRecordService.updateAcademicRecord(id, updatedAcademicRecord);
     }
 
     @DeleteMapping("/{id}")
-    public String deleteAcademicRecord(@PathVariable int id) {
+    public String deleteAcademicRecord(@PathVariable long id) {
         boolean deleted = academicRecordService.deleteAcademicRecord(id);
         return deleted ? "Registro académico eliminado correctamente" : "No se encontró el registro académico con ID: " + id;
     }
