+++ demo/src/main/java/co/edu/udes/activity/backend/demo/services/EnrollmentService.java
@@ -1,7 +1,7 @@
 package co.edu.udes.activity.backend.demo.services;
 
 import co.edu.udes.activity.backend.demo.models.Enrollment;
-import co.edu.udes.activity.backend.demo.repository.EnrollmentRepository;
+import co.edu.udes.activity.backend.demo.repositories.EnrollmentRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -18,7 +18,7 @@ public class EnrollmentService {
         return enrollmentRepository.findAll();
     }
 
-    public Optional<Enrollment> getEnrollmentById(int id) {
+    public Optional<Enrollment> getEnrollmentById(long id) {
         return enrollmentRepository.findById(id);
     }
 
@@ -26,7 +26,7 @@ public class EnrollmentService {
         return enrollmentRepository.save(enrollment);
     }
 
-    public Enrollment updateEnrollment(int id, Enrollment updatedEnrollment) {
+    public Enrollment updateEnrollment(long id, Enrollment updatedEnrollment) {
         return enrollmentRepository.findById(id).map(enrollment -> {
             enrollment.setStudent(updatedEnrollment.getStudent());
             enrollment.setGroup(updatedEnrollment.getGroup());
@@ -34,8 +34,8 @@ public class EnrollmentService {
         }).orElse(null);
     }
 
-    public boolean deleteEnrollment(int id) {
-        if (enrollmentRepository.existsById(id)) {
+    public boolean deleteEnrollment(long id) {
+        if (enrollmentRepository.existsById( id)) {
             enrollmentRepository.deleteById(id);
             return true;
         }
